--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -101,8 +101,24 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           parallel-finished: true
 
-  publish:
+  benchmark:
     needs: finalize
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        id: setup-python
+        with: {python-version: "3.11"}
+      - name: Install ASV
+        run: >-
+          pipx install --python '${{ steps.setup-python.outputs.python-path }}' asv
+      - name: Run benchmark
+        run: >-
+          cd ./benchmarks
+          asv run --python=same
+
+  publish:
+    needs: benchmark
     if: ${{ github.event_name == 'push' && contains(github.ref, 'refs/tags/') }}
     runs-on: ubuntu-latest
     permissions: