--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -104,6 +104,7 @@ jobs:
   benchmark:
     needs: finalize
     runs-on: ubuntu-latest
+    if: ${{ github.event_name != 'push' && ! contains(github.ref, 'refs/tags/') }}
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
@@ -115,7 +116,10 @@ jobs:
       - name: Run benchmark
         run: |-
           cd ./benchmarks
-          asv run --python=same
+          asv check -E existing
+          asv machine --yes
+          asv continuous ${{ github.event.pull_request.base.sha }} ${{ github.sha }}
+          asv compare ${{ github.event.pull_request.base.sha }} ${{ github.sha }} --split --only-changed >> $GITHUB_SUMMARY
 
   publish:
     needs: benchmark