--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -117,7 +117,6 @@ jobs:
         run: |-
           cd ./benchmarks
           asv setup
-          asv check -E existing
           asv machine --yes
           asv continuous ${{ github.event.pull_request.base.sha }} ${{ github.sha }}
           asv compare ${{ github.event.pull_request.base.sha }} ${{ github.sha }} --split --only-changed >> $GITHUB_SUMMARY