--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -112,7 +112,7 @@ jobs:
         with: {python-version: "3.11"}
       - name: Install ASV
         run: |-
-          pipx install --python '${{ steps.setup-python.outputs.python-path }}' asv
+          pipx install --python '${{ steps.setup-python.outputs.python-path }}' asv virtualenv packaging build
       - name: Run benchmark
         run: |-
           cd ./benchmarks