--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -116,8 +116,8 @@ jobs:
       - name: Run benchmark
         run: |-
           cd ./benchmarks
-          asv setup -v
-          asv continuous ${{ github.event.pull_request.base.sha }} ${{ github.sha }}
+          asv setup -v || true
+          asv continuous ${{ github.event.pull_request.base.sha }} ${{ github.sha }} || true
           asv compare ${{ github.event.pull_request.base.sha }} ${{ github.sha }} --split --only-changed >> $GITHUB_SUMMARY
 
   publish: