--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -117,8 +117,8 @@ jobs:
         run: |-
           cd ./benchmarks
           asv setup -v || true
-          asv continuous ${{ github.event.pull_request.base.sha }} ${{ github.event.pull_request.head.sha }} || true
-          asv compare ${{ github.event.pull_request.base.sha }} ${{ github.event.pull_request.head.sha }} --split --only-changed >> $GITHUB_STEP_SUMMARY
+          asv continuous ${{ github.event.pull_request.base.sha }} HEAD || true
+          asv compare ${{ github.event.pull_request.base.sha }} HEAD --split --only-changed >> $GITHUB_STEP_SUMMARY
 
   publish:
     needs: benchmark