--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -107,6 +107,8 @@ jobs:
     if: ${{ github.event_name != 'push' && ! contains(github.ref, 'refs/tags/') }}
     steps:
       - uses: actions/checkout@v4
+        with:
+          depth: 0
       - uses: actions/setup-python@v5
         id: setup-python
         with: {python-version: "3.11"}
@@ -117,8 +119,8 @@ jobs:
         run: |-
           cd ./benchmarks
           asv setup -v || true
-          asv continuous benchmarks HEAD || true
-          asv compare benchmarks HEAD --split --only-changed >> $GITHUB_STEP_SUMMARY
+          asv continuous ${{ github.event.pull_request.base.sha }} HEAD || true
+          asv compare ${{ github.event.pull_request.base.sha }} HEAD --split --only-changed >> $GITHUB_STEP_SUMMARY
 
   publish:
     needs: benchmark