--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -124,7 +124,7 @@ jobs:
           asv setup -v || true
           asv continuous ${{ github.event.pull_request.base.sha }} HEAD || true
           REPORT=$(asv compare ${{ github.event.pull_request.base.sha }} HEAD --split --only-changed)
-          echo ${REPORT} >> $GITHUB_STEP_SUMMARY
+          echo "${REPORT}" >> $GITHUB_STEP_SUMMARY
           echo "report=${REPORT}" >> $GITHUB_OUTPUT
 
       - uses: mshick/add-pr-comment@v2