--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -132,7 +132,8 @@ jobs:
 
       - uses: mshick/add-pr-comment@v2
         with:
-          message: ${{ env.REPORT_RESULT }}
+          message: |-
+            ${{ env.REPORT_RESULT }}
 
   publish:
     needs: benchmark