## What
* Alphabet version with `unique` usage

## Why
* It seems unique is balance of time and memory

All benchmarks:

| Change   | Before [5345c810] <main>       | After [8f3e0e50] <benchmarks>   | Ratio   | Benchmark (Parameter)                          |
|----------|--------------------------------|---------------------------------|---------|------------------------------------------------|
|          | 152                            | 152                             | 1.00    | bench_alphabet.MemSuite.mem_alphabet_s10_0     |
|          | 512                            | 512                             | 1.00    | bench_alphabet.MemSuite.mem_alphabet_s10_1     |
|          | 4.11k                          | 4.11k                           | 1.00    | bench_alphabet.MemSuite.mem_alphabet_s10_2     |
|          | 40.1k                          | 40.1k                           | 1.00    | bench_alphabet.MemSuite.mem_alphabet_s10_3     |
|          | 400k                           | 400k                            | 1.00    | bench_alphabet.MemSuite.mem_alphabet_s10_4     |
|          | failed                         | 4M                              | n/a     | bench_alphabet.MemSuite.mem_alphabet_s10_5     |
|          | 35.9M                          | 35.7M                           | 1.00    | bench_alphabet.MemSuite.peakmem_alphabet_s10_0 |
|          | 35.9M                          | 36.2M                           | 1.01    | bench_alphabet.MemSuite.peakmem_alphabet_s10_1 |
|          | 36.6M                          | 36M                             | 0.98    | bench_alphabet.MemSuite.peakmem_alphabet_s10_2 |
|          | 36.2M                          | 36.7M                           | 1.01    | bench_alphabet.MemSuite.peakmem_alphabet_s10_3 |
| +        | 37.8M                          | 43.2M                           | 1.14    | bench_alphabet.MemSuite.peakmem_alphabet_s10_4 |
|          | failed                         | 94M                             | n/a     | bench_alphabet.MemSuite.peakmem_alphabet_s10_5 |
| +        | 1.86±0.1μs                     | 2.32±0.05μs                     | 1.25    | bench_alphabet.TimeSuite.time_alphabet_s10_0   |
| -        | 25.6±0.7μs                     | 10.9±0.4μs                      | 0.42    | bench_alphabet.TimeSuite.time_alphabet_s10_1   |
| -        | 1.90±0.04ms                    | 92.3±6μs                        | 0.05    | bench_alphabet.TimeSuite.time_alphabet_s10_2   |
| -        | 180±5ms                        | 976±20μs                        | 0.01    | bench_alphabet.TimeSuite.time_alphabet_s10_3   |
|          | 18.4±0.3s                      | 11.0±0.1ms                      | ~0.00   | bench_alphabet.TimeSuite.time_alphabet_s10_4   |
|          | failed                         | 132±1ms                         | n/a     | bench_alphabet.TimeSuite.time_alphabet_s10_5   |